Abstracted file uploader and progress card away from business logic.
Also hoisted "job" state and removed router to make sharing state to pages simpler.

+++ react-frontend/src/pages/upload.tsx
@@ -2,17 +2,28 @@ import axios from 'axios';
 
 import { SingleFileUploader } from '@/components/file-uploader';
 import { useRef } from 'react';
+import { WallerJob } from '@/types';
+import ProgressCard from '@/components/progress-card';
+import { isFileWithPreview } from '@/lib/utils';
 
 interface UploadData {
   id: string;
 }
 
-export default function UploadPage() {
-  // Currently continuously polls API, consider using Server-Sent Events to
-  // avoid continuously opening a connection
+interface JobData {
+  status: string;
+}
+
+export default function UploadPage(props: {
+  job?: WallerJob;
+  setJob: (job: WallerJob | undefined) => void;
+  setIsEditing: (isEditing: boolean) => void;
+}) {
+  // NOTE: Currently continuously polls API, consider using Server-Sent Events
+  // to avoid continuously opening a connection
   const abortControllerRef = useRef<AbortController>(new AbortController());
 
-  async function uploadFile(file: File) {
+  async function uploadImage(file: File) {
     const res = await axios.postForm<UploadData>(
       (import.meta.env.VITE_SERVER_URL as string) + '/jobs',
       {
@@ -23,23 +34,66 @@ export default function UploadPage() {
       }
     );
 
-    return res.data.id;
+    props.setJob({ image: file, id: res.data.id });
   }
 
-  async function cancelJob(id: string) {
+  async function pollJob(): Promise<number> {
+    if (props.job) {
+      return await axios
+        .get<JobData>(
+          (import.meta.env.VITE_SERVER_URL as string) + `/jobs/${props.job.id}`
+        )
+        .then((res) => {
+          switch (res.data.status) {
+            case 'uploading':
+              return 25;
+            case 'queued':
+              return 50;
+            case 'processing':
+              return 75;
+            case 'done':
+              props.setIsEditing(true);
+              return 100;
+            default:
+              return 0;
+          }
+        });
+    }
+    return 0;
+  }
+
+  async function cancelJob() {
     abortControllerRef.current.abort();
     abortControllerRef.current = new AbortController();
-    await axios.delete(
-      (import.meta.env.VITE_SERVER_URL as string) + `/jobs/${id}`,
-      {
-        signal: AbortSignal.timeout(5000)
+    if (props.job) {
+      await axios.delete(
+        (import.meta.env.VITE_SERVER_URL as string) + `/jobs/${props.job.id}`,
+        {
+          signal: AbortSignal.timeout(5000)
+        }
+      );
+
+      if (isFileWithPreview(props.job.image)) {
+        URL.revokeObjectURL(props.job.image.preview);
       }
-    );
+
+      props.setJob(undefined);
+    }
   }
 
   return (
-    <div className=" container flex min-h-screen flex-col place-content-center ">
-      <SingleFileUploader onUpload={uploadFile} onCancel={cancelJob} />
+    <div className="container flex min-h-screen flex-col place-content-center">
+      {props.job ? (
+        <div className="mx-auto w-full max-w-2xl">
+          <ProgressCard
+            file={props.job.image}
+            onPoll={pollJob}
+            onCancel={cancelJob}
+          />
+        </div>
+      ) : (
+        <SingleFileUploader onUpload={uploadImage} />
+      )}
     </div>
   );
 }
